+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @s0/ghcommit
 
+## 1.2.0
+
+### Minor Changes
+
+- a704fb3: Rename repository argument to repo, and deprecate old argument
+- a704fb3: Allow message to be specified as single string
+
 ## 1.1.0
 
 ### Minor Changes